[Feat] 디테일바텀시트 리사이클러 추가. signIn api 변경

+++ travelmaker_android/app/src/main/java/com/gumibom/travelmaker/data/datasource/meeting/MeetingRemoteDataSource.kt
@@ -7,7 +7,7 @@ import retrofit2.Response
 interface MeetingRemoteDataSource {
     suspend fun getMarkerPositions(latitude : Double, longitude : Double, radius : Double) : Response<MutableList<MarkerPositionResponseDTO>>
 
-    suspend fun getPostDetail(id:Int) : Response<MeetingPostDTO>
+    suspend fun getPostDetail(id:Long) : Response<MeetingPostDTO>
